Log_Server/requirements.txt
certifi==2018.1.18
chardet==3.0.4
click==6.7
clickclick==1.2.2
configparser==3.5.0
connexion==1.3
Flask==0.12.2
idna==2.6
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
mysql-connector-python==8.0.11
protobuf==3.5.2.post1
PyYAML==3.12
requests==2.18.4
six==1.11.0
swagger-spec-validator==2.1.0
urllib3==1.22
Werkzeug==0.14.1
